Move some unit tests back into a test project
When breaking up our unit test assemblies I accidentally moved some test code to a non-unit test project.  Split the file up and moved the code back to the correct test project.
