+++ src/EditorFeatures/TestUtilities/ServicesTestUtilities.csproj
@@ -163,7 +163,7 @@
     <Compile Include="Diagnostics\AbstractSuppressionDiagnosticTest.cs" />
     <Compile Include="Diagnostics\AbstractUserDiagnosticTest.cs" />
     <Compile Include="Diagnostics\DiagnosticProviderTestUtilities.cs" />
-    <Compile Include="Diagnostics\DiagnosticsSquiggleTaggerProviderTests.cs" />
+    <Compile Include="Diagnostics\DiagnosticTaggerWrapper.cs" />
     <Compile Include="Diagnostics\GenerateType\GenerateTypeTestState.cs" />
     <Compile Include="Diagnostics\GenerateType\TestGenerateTypeOptionsService.cs" />
     <Compile Include="Diagnostics\GenerateType\TestProjectManagementService.cs" />
